web/dev-requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile dev-requirements.in
#
alembic==1.7.6
    # via
    #   -r requirements.txt
    #   flask-migrate
astroid==2.9.3
    # via pylint
certifi==2021.10.8
    # via
    #   -r requirements.txt
    #   requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.12
    # via
    #   -r requirements.txt
    #   requests
click==8.0.4
    # via
    #   -r requirements.txt
    #   flask
cvxpy==1.1.18
    # via -r requirements.txt
distlib==0.3.4
    # via virtualenv
ecos==2.0.10
    # via
    #   -r requirements.txt
    #   cvxpy
filelock==3.6.0
    # via virtualenv
flake8==4.0.1
    # via -r dev-requirements.in
flask==2.0.3
    # via
    #   -r requirements.txt
    #   flask-migrate
    #   flask-sqlalchemy
flask-migrate==3.1.0
    # via -r requirements.txt
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.txt
    #   flask-migrate
greenlet==1.1.2
    # via
    #   -r requirements.txt
    #   pynvim
    #   sqlalchemy
identify==2.4.11
    # via pre-commit
idna==3.3
    # via
    #   -r requirements.txt
    #   requests
isort==5.10.1
    # via pylint
itsdangerous==2.1.1
    # via
    #   -r requirements.txt
    #   flask
jinja2==3.0.3
    # via
    #   -r requirements.txt
    #   flask
lazy-object-proxy==1.7.1
    # via astroid
mako==1.2.0
    # via
    #   -r requirements.txt
    #   alembic
markupsafe==2.1.0
    # via
    #   -r requirements.txt
    #   jinja2
    #   mako
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
msgpack==1.0.3
    # via pynvim
mypy==0.940
    # via
    #   -r dev-requirements.in
    #   sqlalchemy-stubs
mypy-extensions==0.4.3
    # via mypy
neovim==0.3.1
    # via -r dev-requirements.in
nodeenv==1.6.0
    # via pre-commit
numpy==1.22.2
    # via
    #   -r requirements.txt
    #   cvxpy
    #   ecos
    #   osqp
    #   pandas
    #   qdldl
    #   scipy
    #   scs
osqp==0.6.2.post5
    # via
    #   -r requirements.txt
    #   cvxpy
pandas==1.4.1
    # via -r requirements.txt
platformdirs==2.5.1
    # via
    #   pylint
    #   virtualenv
pre-commit==2.17.0
    # via -r dev-requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.txt
pycodestyle==2.8.0
    # via
    #   -r dev-requirements.in
    #   flake8
pyflakes==2.4.0
    # via flake8
pylint==2.12.2
    # via -r dev-requirements.in
pynvim==0.4.3
    # via neovim
python-dateutil==2.8.2
    # via
    #   -r requirements.txt
    #   pandas
pytz==2021.3
    # via
    #   -r requirements.txt
    #   pandas
pyyaml==6.0
    # via
    #   -r requirements.txt
    #   pre-commit
qdldl==0.1.5.post0
    # via
    #   -r requirements.txt
    #   osqp
requests==2.27.1
    # via -r requirements.txt
scipy==1.8.0
    # via
    #   -r requirements.txt
    #   cvxpy
    #   ecos
    #   osqp
    #   qdldl
    #   scs
scs==3.2.0
    # via
    #   -r requirements.txt
    #   cvxpy
six==1.16.0
    # via
    #   -r requirements.txt
    #   python-dateutil
    #   virtualenv
sqlalchemy==1.4.32
    # via
    #   -r requirements.txt
    #   alembic
    #   flask-sqlalchemy
sqlalchemy-stubs==0.4
    # via -r dev-requirements.in
toml==0.10.2
    # via
    #   pre-commit
    #   pylint
tomli==2.0.1
    # via mypy
types-pyyaml==6.0.4
    # via -r dev-requirements.in
types-requests==2.27.11
    # via -r dev-requirements.in
types-urllib3==1.26.10
    # via types-requests
typing-extensions==4.1.1
    # via
    #   astroid
    #   mypy
    #   pylint
    #   sqlalchemy-stubs
urllib3==1.26.8
    # via
    #   -r requirements.txt
    #   requests
virtualenv==20.13.3
    # via pre-commit
werkzeug==2.0.3
    # via
    #   -r requirements.txt
    #   flask
wrapt==1.13.3
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
